Add files via upload
This is a python script for the android app "QPython3 L", which is a python interpreter for android. The app is free and available in the PlayStore app. With this "dirty code" program you can scan your network for devices without the mac address. A script with mac address scanning will be delivered soon. WARNING: I am a medium Python programmer, who tries to learn and enjoy the Python language. Not just on android but also on Windows and Linux or MacOS I am motivated to learn the behaviour of the Python language. Consequently, this python scanner isn't perfect or finished, because of my limited experiences on android and the restricted system access. (Normally, you would scan your Network with Python Scapy, but hat requieres root access, which I don't expect on android) You can download / copy the code and customize it for your own intentions. I am not responsible for any damages of the code!

* I successfully tested the script unter a lot of different Xiaomi phones. I don't know how other producers phones can run the program correctly.